Middleware topic update for 3.0 (#13968)
