+++ aspnetcore/fundamentals/middleware/index/snapshot/Chain/Startup.cs
@@ -1,29 +1,17 @@
-﻿#define PrimarySample
-#if PrimarySample
-using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Http;
-
-namespace Chain
+﻿public class Startup
 {
-    #region snippet1
-    public class Startup
+    public void Configure(IApplicationBuilder app)
     {
-        public void Configure(IApplicationBuilder app)
+        app.Use(async (context, next) =>
         {
-            app.Use(async (context, next) =>
-            {
-                // Do work that doesn't write to the Response.
-                await next.Invoke();
-                // Do logging or other work that doesn't write to the Response.
-            });
+            // Do work that doesn't write to the Response.
+            await next.Invoke();
+            // Do logging or other work that doesn't write to the Response.
+        });
 
-            app.Run(async context =>
-            {
-                await context.Response.WriteAsync("Hello from 2nd delegate.");
-            });
-        }
+        app.Run(async context =>
+        {
+            await context.Response.WriteAsync("Hello from 2nd delegate.");
+        });
     }
-    #endregion
 }
-#endif
