+++ aspnetcore/fundamentals/middleware/index/snapshot/Chain/StartupMap.cs
@@ -1,11 +1,4 @@
-﻿//#define Map
-#if Map
-using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Http;
-
-#region snippet1
-public class Startup
+﻿public class Startup
 {
     private static void HandleMapTest1(IApplicationBuilder app)
     {
@@ -35,5 +28,3 @@ public class Startup
         });
     }
 }
-#endregion
-#endif
