+++ aspnetcore/fundamentals/middleware/index/snapshot/Middleware/Startup.cs
@@ -1,9 +1,4 @@
-﻿using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Http;
-
-#region snippet1
-public class Startup
+﻿public class Startup
 {
     public void Configure(IApplicationBuilder app)
     {
@@ -13,4 +8,3 @@ public class Startup
         });
     }
 }
-#endregion
