+++ aspnetcore/fundamentals/middleware/write/snapshot/RequestCultureMiddleware.cs
