+++ aspnetcore/fundamentals/middleware/write/snapshot/Startup.cs
@@ -1,0 +1,13 @@
+﻿public class Startup
+{
+    public void Configure(IApplicationBuilder app)
+    {
+        app.UseRequestCulture();
+
+        app.Run(async (context) =>
+        {
+            await context.Response.WriteAsync(
+                $"Hello {CultureInfo.CurrentCulture.DisplayName}");
+        });
+    }
+}
