+++ aspnetcore/fundamentals/middleware/write/snapshot/StartupCulture.cs
@@ -1,0 +1,27 @@
+﻿public class Startup
+{
+    public void Configure(IApplicationBuilder app)
+    {
+        app.Use(async (context, next) =>
+        {
+            var cultureQuery = context.Request.Query["culture"];
+            if (!string.IsNullOrWhiteSpace(cultureQuery))
+            {
+                var culture = new CultureInfo(cultureQuery);
+
+                CultureInfo.CurrentCulture = culture;
+                CultureInfo.CurrentUICulture = culture;
+            }
+
+            // Call the next delegate/middleware in the pipeline
+            await next();
+        });
+
+        app.Run(async (context) =>
+        {
+            await context.Response.WriteAsync(
+                $"Hello {CultureInfo.CurrentCulture.DisplayName}");
+        });
+
+    }
+}
